systemtools/managecert/requirements.txt
pyOpenSSL==20.0.1